requirements.txt
python-gitlab
argparse
urllib3
requests